+++ App.tsx
@@ -4,7 +4,7 @@ import { SafeAreaProvider } from 'react-native-safe-area-context';
 
 import useCachedResources from './hooks/useCachedResources';
 import useColorScheme from './hooks/useColorScheme';
-import Navigation from './navigation';
+import Navigation from './src/navigation';
 import { Ionicons } from '@expo/vector-icons';
 
 export default function App() {
