+++ src/components/QuoteCard.tsx
