+++ src/components/Themed.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Text as DefaultText, View as DefaultView } from 'react-native';
 
 import Colors from '../constants/Colors';
-import useColorScheme from '../hooks/useColorScheme';
+import useColorScheme from '../../hooks/useColorScheme';
 
 export function useThemeColor(
   props: { light?: string; dark?: string },
