+++ src/screens/TabOneScreen.tsx
@@ -123,5 +123,5 @@ const styles = StyleSheet.create({
     height: 50,
     margin: 10
     // flex: 1
-  }
+  },
 });
